requirements.txt
asgiref==3.4.1
Django==3.2.7
pkg-resources==0.0.0
psycopg2==2.9.1
pytz==2021.1
RPi.GPIO
sqlparse==0.4.2
typing-extensions==3.10.0.2
